# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.APIErrorResponse:
    properties:
      code:
        type: integer
      message: {}
    type: object
  handlers.APIResponse:
    properties:
      code:
        type: integer
      message: {}
    type: object
...
  sys.StatsResponse:
    properties:
      cpu:
        $ref: '#/definitions/system.CPUInfo'
      load:
        $ref: '#/definitions/system.AverageLoad'
      memory:
        $ref: '#/definitions/system.RAMInfo'
      temperature:
        items:
          $ref: '#/definitions/system.SensorInfo'
        type: array
    type: object
  system.AverageLoad:
    properties:
      load1:
        type: number
      load5:
        type: number
      load15:
        type: number
    type: object
  system.CPUInfo:
    properties:
      cache:
        type: integer
      coreCount:
        type: integer
      cores:
        items:
          $ref: '#/definitions/system.CpuCore'
        type: array
      load:
        type: number
      model:
        type: string
      name:
        type: string
      threadCount:
        type: integer
    type: object
  system.CpuCore:
    properties:
      frequency:
        type: number
      id:
        type: string
    type: object
  system.HostInfo:
    properties:
      addrs:
        items:
          type: string
        type: array
      hostname:
        type: string
      os:
        type: string
      procs:
        type: integer
      uptime:
        type: number
      virt:
        type: string
    type: object
  system.RAMInfo:
    properties:
      available:
        type: integer
      total:
        type: integer
      use:
        type: number
      used:
        type: integer
    type: object
  system.SensorInfo:
    properties:
      current:
        type: number
      key:
        type: string
      max:
        type: number
    type: object
  system.SystemdUnit:
    properties:
      preset:
        type: string
      state:
        type: string
      unit_file:
        type: string
    type: object
host: 0.0.0.0:3000
info:
  contact: {}
  title: Desky API
  version: "1.0"
paths:
...
  /system/info:
    get:
      description: host information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.HostInfo'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIErrorResponse'
      summary: HostInfo
      tags:
      - system
  /system/stats:
    get:
      description: host information ws interval update = 5s
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sys.StatsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIErrorResponse'
      summary: HostStatsWS
      tags:
      - system
  /system/systemd/{service}/{command}:
    post:
      description: execute device command
      parameters:
      - description: systemd command
        in: path
        name: command
        required: true
        type: string
      - description: systemd service
        in: path
        name: service
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/handlers.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIErrorResponse'
        "501":
          description: Uninplemented
          schema:
            $ref: '#/definitions/handlers.APIErrorResponse'
      summary: UnitCommand
      tags:
      - system
  /system/systemd/status:
    get:
      description: units systemd list
      parameters:
      - description: Page number for pagination (optional)
        in: query
        name: page
        type: string
      - description: Number of items per page (optional)
        in: query
        name: count
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/system.SystemdUnit'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIErrorResponse'
      summary: SystemdUnits
      tags:
      - system
swagger: "2.0"